+++ packages/react-vapor/src/components/editor/CodeEditor.tsx
@@ -1,4 +1,5 @@
 import 'codemirror/addon/dialog/dialog';
+import 'codemirror/addon/fold/foldgutter';
 import 'codemirror/addon/hint/anyword-hint';
 import 'codemirror/addon/hint/show-hint';
 import 'codemirror/addon/search/jump-to-line';