+++ packages/react-vapor/src/components/form/index.ts
@@ -1,2 +1,3 @@
 export * from './Form';
 export * from './FormGroup';
+export {FormProvider} from '../form/FormProvider';