+++ packages/react-vapor/src/components/headers/tests/TabsHeader.spec.tsx
@@ -2,11 +2,11 @@ import {mount, ReactWrapper, shallow} from 'enzyme';
 import * as React from 'react';
 import {Provider} from 'react-redux';
 import {Store} from 'redux';
+
 import {IReactVaporState} from '../../../ReactVaporState';
 import {clearState} from '../../../utils/ReduxUtils';
 import {TestUtils} from '../../../utils/tests/TestUtils';
-import {ITabProps} from '../../tab/Tab';
-import {TabConnected} from '../../tab/TabConnected';
+import {ITabProps, TabConnected} from '../../tab/Tab';
 import {TabNavigation} from '../../tab/TabNavigation';
 import {ITabsHeaderProps, TabsHeader} from '../TabsHeader';