+++ packages/react-vapor/src/components/tab/TabContent.tsx
@@ -3,12 +3,8 @@ import * as React from 'react';
 
 export interface ITabContentProps extends React.HTMLAttributes<HTMLDivElement> {}
 
-export class TabContent extends React.Component<ITabContentProps> {
-    render() {
-        return (
-            <div {...this.props} className={classNames('tab-content', this.props.className)}>
-                {this.props.children}
-            </div>
-        );
-    }
-}
+export const TabContent: React.FunctionComponent<ITabContentProps> = ({className, children, ...divProps}) => (
+    <div {...divProps} className={classNames('tab-content', className)}>
+        {children}
+    </div>
+);