+++ packages/react-vapor/src/components/table-hoc/tests/TableHeaderWithSort.spec.tsx
@@ -1,6 +1,8 @@
 import {shallowWithState, shallowWithStore} from '@helpers/enzyme-redux';
 import * as React from 'react';
+import {render, screen} from '@test-utils';
 
+import userEvent from '@testing-library/user-event';
 import {getStoreMock, ReactVaporMockStore} from '../../../utils/tests/TestUtils';
 import {TextLoadingPlaceholder} from '../../loading';
 import {TableHeaderActions} from '../actions/TableHeaderActions';
@@ -47,7 +49,7 @@ describe('Table HOC', () => {
         });
 
         it('should not throw when rendering children', () => {
-            const render = () =>
+            const component = () =>
                 shallowWithStore(
                     <TableHeaderWithSort {...defaultProps}>
                         <div>Hello</div>
@@ -57,7 +59,7 @@ describe('Table HOC', () => {
                     .dive()
                     .dive();
 
-            expect(render).not.toThrow();
+            expect(component).not.toThrow();
         });
 
         it('should dispatch an addTableHeader on componentDidMount', () => {
@@ -98,6 +100,33 @@ describe('Table HOC', () => {
                 .dive();
 
             expect(wrapper.find(TextLoadingPlaceholder).length).toBe(1);
+        });
+
+        it('renders the default svg when the table is not sorted', () => {
+            render(<TableHeaderWithSort id={'patate'} tableId={'id'} isLoading={false}></TableHeaderWithSort>);
+
+            expect(screen.getByRole('img', {name: /asc-desc icon/i})).toBeInTheDocument();
+        });
+
+        it('renders the sorted-asc svg when the user click one time on the arrows', () => {
+            render(<TableHeaderWithSort id={'patate'} tableId={'id'} isLoading={false}></TableHeaderWithSort>);
+
+            const btn = screen.getByRole('img', {name: /asc-desc icon/i});
+            userEvent.click(btn);
+
+            expect(screen.getByRole('img', {name: /sorted-asc icon/i})).toBeInTheDocument();
+        });
+
+        it('renders the sorted-desc svg when the user click two time on the arrows', () => {
+            render(<TableHeaderWithSort id={'patate'} tableId={'id'} isLoading={false}></TableHeaderWithSort>);
+
+            const btn1 = screen.getByRole('img', {name: /asc-desc icon/i});
+            userEvent.click(btn1);
+
+            const btn2 = screen.getByRole('img', {name: /sorted-asc icon/i});
+            userEvent.click(btn2);
+
+            expect(screen.getByRole('img', {name: /sorted-desc icon/i})).toBeInTheDocument();
         });
     });
 });