+++ packages/react-vapor/src/components/toast/tests/Toast.spec.tsx
@@ -1,4 +1,4 @@
-import {render, screen, waitFor} from '@testing-library/react';
+import {render, screen, waitFor} from '@test-utils';
 import userEvent from '@testing-library/user-event';
 import {mount, ReactWrapper, shallow} from 'enzyme';
 import * as React from 'react';